errors fixed

+++ app/src/main/java/com/unava/dia/weatherforecast/data/model/future/Forecastday.kt
@@ -2,7 +2,7 @@ package com.unava.dia.weatherforecast.data.model.future
 
 data class Forecastday (
     var date : String? = null,
-    var date_epoch : Int? = null,
+    var date_epoch : Long? = null,
     var day : Day? = null,
     var astro : Astro? = null,
     var hour : List<Hour>? = null
